pyproject.toml
[tool.poetry]
name = "pytest-mini"
version = "0.1.11"
description = "A plugin to test mp"
authors = ["听白 <490336534@qq.com>"]
readme = "README.md"
packages = [{ include = "pytest_mini/" }]

[tool.poetry.dependencies]
python = "^3.9"
allure-pytest = "^2.11.1"
allure-python-commons = "^2.11.1"
attrs = "^22.1.0"
certifi = "^2022.9.24"
charset-normalizer = "^3.0.1"
concurrent-log-handler = "^0.9.20"
ddt = "^1.6.0"
exceptiongroup = "^1.0.4"
Faker = "^15.3.3"
future = "^0.18.2"
idna = "^3.4"
iniconfig = "^2.0.0"
minium = "^1.3.1"
packaging = "^23.0"
pluggy = "^1.0.0"
portalocker = "^2.6.0"
pyee = "^9.0.4"
pyparsing = "^3.0.9"
pytest = "^7.2.0"
pytest-assume = "^2.4.3"
pytest-dependency = "^0.5.1"
pytest-lazy-fixture = "^0.6.3"
pytest-ordering = "^0.6"
python-dateutil = "^2.8.2"
PyYAML = "^6.0"
requests = "^2.28.1"
six = "^1.16.0"
tomli = "^2.0.1"
typing_extensions = "^4.4.0"
urllib3 = "^1.26.12"

[tool.poetry.dev-dependencies]
twine = "^4.0.2"


[[tool.poetry.source]]
name = "tsinghua"
default = true
url = "https://pypi.tuna.tsinghua.edu.cn/simple"


[tool.poetry.plugins.pytest11]
mini = "pytest_mini.mini"
mp = "pytest_mini.plugin:Plugin"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
